+++ thunderstore/WorkbenchesPlus/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.0.14
+
+### UI IMPROVEMENTS
+- Build range cycle is absolute: default, then 50, 100, 150, 200 m (not added on top of vanilla)
+
+## 1.0.13
+
+### FIX
+- Opening a chest no longer crashes when category chips were already destroyed
+
 ## 1.0.12
 
 ### UI IMPROVEMENTS